+++ recipes/default.rb
@@ -1,13 +1,13 @@
 #
-# Cookbook Name:: consul-template
+# Cookbook:: consul-template
 # Recipe:: default
 #
-# Copyright (C) 2014
+# Copyright:: (C) 2014
 #
 #
 #
 
-if node['platform'] == 'windows'
+if platform?('windows')
   include_recipe "consul-template::install_windows_#{node['consul_template']['install_method']}"
   include_recipe "consul-template::service_windows"
 else