+++ recipes/install_binary.rb
@@ -1,11 +1,13 @@
 #
-# Cookbook Name:: consul-template
+# Cookbook:: consul-template
 # Recipe:: install_binary
 #
-# Copyright (C) 2014
+# Copyright:: (C) 2014
 #
 #
 #
+
+package 'unzip'
 
 remote_file "#{Chef::Config[:file_cache_path]}/consul-template-#{node['consul_template']['version']}.zip" do
   source "https://releases.hashicorp.com/consul-template/#{node['consul_template']['version']}/consul-template_#{node['consul_template']['version']}_linux_amd64.zip"
@@ -19,9 +21,4 @@ bash 'extract consul-template' do
   code <<~EOH
     unzip -o "#{Chef::Config[:file_cache_path]}/consul-template-#{node['consul_template']['version']}.zip" -d /usr/local/bin
   EOH
-  if node['consul_template']['init_style'] == 'runit'
-    notifies :restart, 'runit_service[consul-template]', :delayed
-  else
-    notifies :restart, 'service[consul-template]', :delayed
-  end
 end